refactor : BottomSheetItem 공통으로 추출
